+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## Version 1.3.0
+
+- Use `vscode.env.appRoot` to get the app folder path
+- Self-elevation support when write access is denied
+
+## Version 1.2.2
+
+- Use a more reliable way to get the app folder path
+
+## Version 1.2.1
+
+- Fix for VSCode 1.93.0 where breaking the way to get the app folder path
+
+## Version 1.2.0
+
+- Fix for VSCode 1.86.0 and later
+
 ## Version 1.1.0
 
 - Only create backup file if none exists yet for current VSCode version